// src/generated/api.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/api/v1/prefectures": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description 都道府県に関する一覧データを返します。 */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PrefecturesResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/population/composition/perYear": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description 地域単位、年単位の年齢構成のデータを返します。 */
        get: {
            parameters: {
                query: {
                    prefCode: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PopulationCompositionPerYearResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        Prefecture: {
            /**
             * @description 都道府県コード
             * @example 1
             */
            prefCode: number;
            /**
             * @description 都道府県名
             * @example 北海道
             */
            prefName: string;
        };
        PrefecturesResponse: {
            message: unknown;
            result: components["schemas"]["Prefecture"][];
        };
        PopulationCompositionPerYear: {
            /**
             * @description 実績値と推計値の区切り年
             * @example 2020
             */
            boundaryYear: number;
            data: {
                /**
                 * @description ラベル
                 * @example 年少人口
                 */
                label: string;
                data: {
                    /**
                     * @description 年
                     * @example 1960
                     */
                    year: number;
                    /**
                     * @description 人口
                     * @example 1681479
                     */
                    value: number;
                    /**
                     * @description 割合（label が総人口の場合は存在しません）
                     * @example 33.37
                     */
                    rate?: number;
                }[];
            }[];
        };
        PopulationCompositionPerYearResponse: {
            message: unknown;
            result: components["schemas"]["PopulationCompositionPerYear"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;